add phashed and hammel dist

+++ phashml.py
@@ -4,19 +4,27 @@ from pyphashml.phashml import phashml_distance
 import os
 import itertools
 import pandas as pd
+from bitstring import BitArray
+
+
+def hdist(a, b):
+    return str(a**b).count(1)
+
+
+print(phashml_distance(phashmlctx.image_hash("./images/" + "0044500073225_1.400.jpg"),
+      phashmlctx.image_hash("./images/" + "0017082876362_2.400.jpg")))
 
 im = os.listdir("./images/")
+if os.path.isfile("./hashes.csv"):
+    hashes = pd.read_csv("./hashes.csv")['hashes']
+    hashes = [h for h in hashes[:1000]]
+else:
 
-hashes = [phashmlctx.image_hash("./images/" + ima) for ima in im]
-pd.DataFrame({"file": im, "hashes": hashes}).to_csv('hashes.csv')
-print((len(im)**2 - len(im)) / 2)
-df = pd.DataFrame({"x": [], "y": [], "d": []})
-for n, (x, y) in enumerate(itertools.combinations(hashes, 2)):
+    hashes = [phashmlctx.image_hash("./images/" + ima) for ima in im]
+    pd.DataFrame({"file": im, "hashes": hashes}).to_csv('hashes.csv')
 
-    d = (phashml_distance(x, y))
-    df.append({"x": x, "y": y, "d": d}, ignore_index=True)
 
-    if n % 50 == 0:
-        print(n)
+dist = [[x, y, phashml_distance(BitArray(x), BitArray(y))]
+        for x, y in itertools.combinations(hashes, 2)]
 
-df.to_csv("phashmldist.csv")
+pd.DataFrame(dist).to_csv("phashmldist.csv")
